# core/docs/swagger.yaml
definitions:
...
  controller.GroupMCPResponse:
    properties:
      created_at:
        type: string
      endpoints:
        $ref: '#/definitions/controller.MCPEndpoint'
      group_id:
        type: string
      id:
        type: string
      name:
        type: string
      openapi_config:
        $ref: '#/definitions/model.MCPOpenAPIConfig'
      proxy_config:
        $ref: '#/definitions/model.GroupMCPProxyConfig'
      status:
        $ref: '#/definitions/model.GroupMCPStatus'
      type:
        $ref: '#/definitions/model.GroupMCPType'
      update_at:
        type: string
    type: object
...
  controller.MCPEndpoint:
    properties:
      host:
        type: string
      sse:
        type: string
      streamable_http:
        type: string
    type: object
...
  controller.UpdateGroupMCPStatusRequest:
    properties:
      status:
        $ref: '#/definitions/model.GroupMCPStatus'
    type: object
...
  middleware.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.GroupMCP:
    properties:
      created_at:
        type: string
      group_id:
        type: string
      id:
        type: string
      name:
        type: string
      openapi_config:
        $ref: '#/definitions/model.MCPOpenAPIConfig'
      proxy_config:
        $ref: '#/definitions/model.GroupMCPProxyConfig'
      status:
        $ref: '#/definitions/model.GroupMCPStatus'
      type:
        $ref: '#/definitions/model.GroupMCPType'
      update_at:
        type: string
    type: object
  model.GroupMCPProxyConfig:
    properties:
      headers:
        additionalProperties:
          type: string
        type: object
      querys:
        additionalProperties:
          type: string
        type: object
      url:
        type: string
    type: object
  model.GroupMCPStatus:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - GroupMCPStatusEnabled
    - GroupMCPStatusDisabled
  model.GroupMCPType:
    enum:
    - mcp_proxy_sse
    - mcp_proxy_streamable
    - mcp_openapi
    type: string
    x-enum-varnames:
    - GroupMCPTypeProxySSE
    - GroupMCPTypeProxyStreamable
    - GroupMCPTypeOpenAPI
...
  model.MCPOpenAPIConfig:
    properties:
      authorization:
        type: string
      openapi_content:
        type: string
      openapi_spec:
        type: string
      server_addr:
        type: string
      v2:
        type: boolean
    type: object
...
  openai.SubscriptionResponse:
    properties:
      access_until:
        type: integer
      hard_limit_usd:
        type: number
      has_payment_method:
        type: boolean
      object:
        type: string
      soft_limit_usd:
        type: number
      system_hard_limit_usd:
        type: number
    type: object
  openai.UsageResponse:
    properties:
      object:
        type: string
      total_usage:
        description: DailyCosts []OpenAIUsageDailyCost `json:"daily_costs"`
        type: number
    type: object
info:
  contact: {}
  title: AI Proxy Swagger API
  version: "1.0"
paths:
...
  /api/mcp/group/{group}:
    get:
      description: Get a list of Group MCPs with pagination and filtering
      parameters:
      - description: Group ID
        in: path
        name: group
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: MCP type
        in: query
        name: type
        type: string
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: MCP status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controller.GroupMCPResponse'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Group MCPs
      tags:
      - mcp
    post:
      consumes:
      - application/json
      description: Create a new Group MCP
      parameters:
      - description: Group ID
        in: path
        name: group
        required: true
        type: string
      - description: Group MCP object
        in: body
        name: mcp
        required: true
        schema:
          $ref: '#/definitions/model.GroupMCP'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.GroupMCPResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create Group MCP
      tags:
      - mcp
  /api/mcp/group/{group}/{id}:
    delete:
      description: Delete a Group MCP by ID and Group ID
      parameters:
      - description: MCP ID
        in: path
        name: id
        required: true
        type: string
      - description: Group ID
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Group MCP
      tags:
      - mcp
    get:
      description: Get a specific Group MCP by its ID and Group ID
      parameters:
      - description: MCP ID
        in: path
        name: id
        required: true
        type: string
      - description: Group ID
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.GroupMCPResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Group MCP by ID
      tags:
      - mcp
    put:
      consumes:
      - application/json
      description: Update an existing Group MCP
      parameters:
      - description: MCP ID
        in: path
        name: id
        required: true
        type: string
      - description: Group ID
        in: path
        name: group
        required: true
        type: string
      - description: Group MCP object
        in: body
        name: mcp
        required: true
        schema:
          $ref: '#/definitions/model.GroupMCP'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.GroupMCPResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update Group MCP
      tags:
      - mcp
  /api/mcp/group/{group}/{id}/status:
    post:
      consumes:
      - application/json
      description: Update the status of a Group MCP
      parameters:
      - description: MCP ID
        in: path
        name: id
        required: true
        type: string
      - description: Group ID
        in: path
        name: group
        required: true
        type: string
      - description: MCP status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateGroupMCPStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Group MCP status
      tags:
      - mcp
  /api/mcp/group/all:
    get:
      description: Get all Group MCPs with filtering
      parameters:
      - description: MCP status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controller.GroupMCPResponse'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all Group MCPs
      tags:
      - mcp
...
  /mcp:
    delete:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Host MCP Streamable Server
    get:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Host MCP Streamable Server
    post:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Host MCP Streamable Server
  /mcp/group/{id}:
    delete:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Group MCP Streamable Server
    get:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Group MCP Streamable Server
    post:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Group MCP Streamable Server
  /mcp/group/{id}/sse:
    get:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Group MCP SSE Server
  /mcp/group/message:
    post:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: MCP SSE Proxy
...
  /sse:
    get:
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Public MCP SSE Server
...
  /v1/dashboard/billing/subscription:
    get:
      description: Get subscription
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.SubscriptionResponse'
      security:
      - ApiKeyAuth: []
      summary: Get subscription
      tags:
      - relay
  /v1/dashboard/billing/usage:
    get:
      description: Get usage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.UsageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get usage
      tags:
      - relay
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"